profile and home fragment ui update
