+++ babyOne/app/src/main/AndroidManifest.xml
@@ -14,11 +14,11 @@
         android:theme="@style/Theme.BabyOne"
         tools:targetApi="31">
         <activity
-            android:name=".Medicine"
+            android:name=".BmiGraph"
             android:exported="false"
             android:screenOrientation="sensorPortrait" />
         <activity
-            android:name=".VaccineUpdate"
+            android:name=".WeightGraph"
             android:exported="false"
             android:screenOrientation="sensorPortrait" />
         <activity
@@ -26,7 +26,7 @@
             android:exported="true"
             android:screenOrientation="sensorPortrait" />
         <activity
-            android:name=".BmiToGraph"
+            android:name=".HeightGraph"
             android:exported="false"
             android:screenOrientation="sensorPortrait" />
         <activity
