+++ babyOne/app/src/main/java/com/example/babyone/HeightGraph.java
@@ -1,0 +1,34 @@
+package com.example.babyone;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+public class HeightGraph extends AppCompatActivity {
+
+    Button btnback;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_height_graph);
+
+        btnback = findViewById(R.id.btnBack);
+
+        /*ACTION WHICH OCCURS WHEN THE BACK BUTTON IS CLICKED*/
+        btnback.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                finish();
+            }
+        });
+    }
+
+
+
+
+}
