+++ babyOne/app/src/main/java/com/example/babyone/WeightGraph.java
@@ -1,0 +1,28 @@
+package com.example.babyone;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+public class WeightGraph extends AppCompatActivity {
+
+    Button btnback_2;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_weight_graph);
+
+        btnback_2 = findViewById(R.id.btnBack_vaccine);
+
+        /*ACTION WHICH OCCURS WHEN THE BACK BUTTON IS CLICKED*/
+        btnback_2.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                finish();
+            }
+        });
+    }
+}
