+++ babyOne/app/src/main/res/layout/activity_height_graph.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BmiToGraph">
+    tools:context=".HeightGraph">
 
     <Button
         android:id="@+id/btnBack"
